Finished moving traits
