+++ backend/libs/rulesets/sign/Cargo.toml
@@ -7,3 +7,5 @@ edition = "2018"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+bytes = "1.0.1"
+ring = "0.16.20"
