+++ backend/libs/rulesets/trade_observer/Cargo.toml
@@ -7,3 +7,9 @@ edition = "2018"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+types = { path = "../../types" }
+entities = { path = "../entities" }
+
+chrono = "0.4.19"
+futures-core = "0.3.12"
+async-trait = "0.1.42"
