+++ backend/libs/rulesets/user_stream/Cargo.toml
@@ -7,3 +7,7 @@ edition = "2018"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+types = { path = "../../types" }
+entities = { path = "../entities" }
+
+async-trait = "0.1.42"
