+++ backend/libs/rulesets/user_stream/src/lib.rs
@@ -1,7 +1,15 @@
-#[cfg(test)]
-mod tests {
-  #[test]
-  fn it_works() {
-    assert_eq!(2 + 2, 4);
-  }
+use ::async_trait::async_trait;
+
+use ::entities::APIKeyInner;
+use ::types::GenericResult;
+
+#[async_trait]
+pub trait UserStream {
+  async fn get_listen_key(&self, api_key: &APIKeyInner) -> GenericResult<()>;
+  async fn clise_listen_key(
+    &self,
+    api_key: &APIKeyInner,
+    listen_key: &String,
+  ) -> GenericResult<()>;
+  async fn start(&self) -> GenericResult<()>;
 }
